Sync merge replace-setup-expt into EXP-port2Gaea and then copied to EXP-port2Gaea_sync

+++ sorc/link_fv3gfs.sh
@@ -2,16 +2,22 @@
 set -ex
 
 #--make symbolic links for EMC installation and hardcopies for NCO delivery
+. ./machine-setup.sh
+echo "target system (machine) set to $target"
 
-RUN_ENVIR=${1:-emc}
-machine=${2:-cray}
-
-if [ $RUN_ENVIR != emc -a $RUN_ENVIR != nco ]; then
-    echo 'Syntax: link_fv3gfs.sh ( nco | emc ) ( cray | theia )'
+if [ -z $target ]; then
+    echo 'target value not set (unknown system not supported)'
     exit 1
 fi
-if [ $machine != cray -a $machine != theia -a $machine != gaea ]; then
-    echo 'Syntax: link_fv3gfs.sh ( nco | emc ) ( cray | theia )'
+
+RUN_ENVIR=${1:-emc}
+
+if [ $RUN_ENVIR != emc -a $RUN_ENVIR != nco ]; then
+    echo 'Syntax: link_fv3gfs.sh ( nco | emc )'
+    exit 1
+fi
+if [ $target != wcoss_cray -a $target != theia -a $target != gaea -a $target != jet ]; then
+    echo '$target value set to unknown or unsupported system'
     exit 1
 fi
 
@@ -21,18 +27,32 @@ LINK="ln -fs"
 pwd=$(pwd -P)
 
 #--model fix fields
-if [ $machine == "cray" ]; then
+if [ $target == "wcoss_cray" ]; then
     FIX_DIR="/gpfs/hps3/emc/global/noscrub/emc.glopara/git/fv3gfs/fix"
-elif [ $machine = "theia" ]; then
+elif [ $target == "theia" ]; then
     FIX_DIR="/scratch4/NCEPDEV/global/save/glopara/git/fv3gfs/fix"
-elif [ $machine = "gaea" ]; then
+elif [ $target == "gaea" ]; then
     FIX_DIR="/lustre/f1/pdata/ncep_shared/FV3GFS_V1_RELEASE/fix"
+elif [ $target == "jet" ]; then
+    FIX_DIR="/lfs3/projects/hfv3gfs/glopara/git/fv3gfs/fix"
+else
+    echo 'CRITICAL: links to fix files not set'
+    exit 1
 fi
-cd ${pwd}/../fix                ||exit 8
-for dir in fix_am fix_fv3 fix_orog fix_fv3_gmted2010 ; do
-    [[ -d $dir ]] && rm -rf $dir
-done
-$LINK $FIX_DIR/* .
+
+if [ ! -r $FIX_DIR ]; then
+   echo "CRITICAL: you do not of read permissions to the location of the fix file $FIX_DIR"
+   exit -1
+fi
+
+if [ ! -z $FIX_DIR ]; then
+ if [ ! -d ${pwd}/../fix ]; then mkdir ${pwd}/../fix; fi
+ cd ${pwd}/../fix                ||exit 8
+ for dir in fix_am fix_fv3 fix_orog fix_fv3_gmted2010 ; do
+     [[ -d $dir ]] && rm -rf $dir
+ done
+ $LINK $FIX_DIR/* .
+fi
 
 
 #--add gfs_post file
@@ -50,7 +70,6 @@ cd ${pwd}/../ush                ||exit 8
     for file in fv3gfs_downstream_nems.sh  fv3gfs_dwn_nems.sh  gfs_nceppost.sh  gfs_transfer.sh  link_crtm_fix.sh  trim_rh.sh fix_precip.sh; do
         $LINK ../sorc/gfs_post.fd/ush/$file                  .
     done
-
 
 #--add GSI/EnKF file
 cd ${pwd}/../jobs               ||exit 8
@@ -72,7 +91,6 @@ cd ${pwd}/../scripts            ||exit 8
 cd ${pwd}/../fix                ||exit 8
     [[ -d fix_gsi ]] && rm -rf fix_gsi
     $LINK ../sorc/gsi.fd/fix  fix_gsi
-
 
 #--add DA Monitor file (NOTE: ensure to use correct version)
 cd ${pwd}/../fix                ||exit 8
@@ -117,10 +135,10 @@ cd ${pwd}/../ush                ||exit 8
     $LINK ../sorc/gsi.fd/util/Radiance_Monitor/nwprod/radmon_shared.v2.0.4/ush/radmon_verf_bcor.sh           .
     $LINK ../sorc/gsi.fd/util/Radiance_Monitor/nwprod/radmon_shared.v2.0.4/ush/radmon_verf_time.sh           .
     
-
 #--link executables 
 
 cd $pwd/../exec
+
 [[ -s fv3_gfs_nh.prod.32bit.x ]] && rm -f fv3_gfs_nh.prod.32bit.x
 $LINK ../sorc/fv3gfs.fd/NEMS/exe/fv3_gfs_nh.prod.32bit.x .
 
@@ -132,5 +150,20 @@ for gsiexe in  global_gsi global_enkf calc_increment_ens.x  getsfcensmeanp.x  ge
     [[ -s $gsiexe ]] && rm -f $gsiexe
     $LINK ../sorc/gsi.fd/exec/$gsiexe .
 done
+
+if [[ $target == "gaea" ]]; then
+  if [[ -f /lustre/f1/pdata/ncep_shared/exec/wgrib2 ]]; then
+   cp /lustre/f1/pdata/ncep_shared/exec/wgrib2 .
+  else
+   echo 'WARNING wgrib2 did not copy from /lustre/f1/pdata/ncep_shared/exec on Gaea'
+  fi
+fi
+if [[ $target == "jet" ]]; then
+  if [[ -f /mnt/lfs3/projects/hfv3gfs/gwv/fv3/exec/wgrib2 ]]; then
+   cp /mnt/lfs3/projects/hfv3gfs/gwv/fv3/exec/wgrib2 .
+  else
+   echo 'WARNING wgrib2 did not copy from /mnt/lfs3/projects/hfv3gfs/gwv/fv3/exec on Jet'
+  fi
+fi
 
 exit 0
